requirements.txt
pydantic==2.10.1
phidata
yfinance
duckduckgo-search
chainlit
langchain
langchain-groq